Gemini: server (fly.io, non-functional)
This puts the framework for building a statically compiled agate into a scratch Docker image which can be run from fly.io (as well as the fly.toml that sets it up correctly).

This doesn't seem to work on fly.io itself at the moment, so more tinkering soon, but it works locally.